refactor: move non route to common
