+++ src/common/paddle/cores/eventHandler.ts
@@ -1,9 +1,9 @@
 import { EventName, type EventEntity } from '@paddle/paddle-node-sdk';
 
-import { SubscriptionProvider } from '../../../../entity/user/User';
-import { logger } from '../../../../logger';
-import { logPaddleAnalyticsEvent } from '../../../../common/paddle';
-import { AnalyticsEventName } from '../../../../integrations/analytics';
+import { SubscriptionProvider } from '../../../entity/user/User';
+import { logger } from '../../../logger';
+import { logPaddleAnalyticsEvent } from '../index';
+import { AnalyticsEventName } from '../../../integrations/analytics';
 import {
   processCoresTransactionCompleted,
   processCoresTransactionCreated,
